fix(browser): prevent hierarchy re-rendering on other focus changes
